Clean up repo and allow returning a result to the browser
