+++ .gitignore
@@ -5,3 +5,5 @@ gpt-settings.talon
 .mypy_cache/
 .pytest_cache/
 __pycache__/
+
+node_modules/
