+++ GPT/beta-commands/function-calling/readme.md
@@ -1,0 +1,3 @@
+# Function Calling for Natural Language
+
+A long term goal of this repository is to allow users to speak to Talon with natural language and have Talon intelligently choose the proper function to perform, even if it doesn't match the exact grammar. Before this can be accomplished, we need stable JSON serializations of each function in community. In addition to a stable and semantically meaningful function name and docstring, we also need a description of what each arg does.
