+++ GPT/modelInsertionMethod.talon-list
@@ -2,7 +2,17 @@ list:  user.modelInsertionMethod
 
 -
 
+# Paste above the current selection
 above
+
+# Past below the current selection
 below
+
+# Instead of pasting, return the result to the clipboard
 clipped
+
+# After pasting, select the text
 selected
+
+# Open the result in a new browser tab
+browser
