+++ docs/usage-examples/bullets.gif
